pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.2.0"
description = ""
authors = ["Vadim Gusak <gusakvd@gmail.com>"]
readme = "README.md"
packages = [{include = "gendiff"}]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
pytest-cov = "^3.0.0"
pytest = "^7.1.3"
coverage = "^6.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
parse = "gendiff.parser:main"
